already sort with negatives doesnt work

+++ libft/srcs/ft_atoi.c
@@ -6,16 +6,16 @@
 /*   By: anremiki <anremiki@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/09/13 01:24:23 by anremiki          #+#    #+#             */
-/*   Updated: 2021/10/14 05:24:09 by anremiki         ###   ########.fr       */
+/*   Updated: 2022/03/14 18:34:44 by anremiki         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_atoi(const char *nptr)
+long	ft_atoi(const char *nptr)
 {
 	int			sign;
-	long int	res;
+	long res;
 
 	res = 0;
 	sign = 1;
