+++ libft/srcs/libft.h
@@ -6,7 +6,7 @@
 /*   By: anremiki <anremiki@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/09/13 00:34:47 by anremiki          #+#    #+#             */
-/*   Updated: 2021/10/24 03:48:29 by anremiki         ###   ########.fr       */
+/*   Updated: 2022/03/14 18:35:16 by anremiki         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -52,7 +52,7 @@ char	*ft_strchr(const char *s, int c);
 void	*ft_calloc(size_t nmemb, size_t size);
 char	*ft_strdup(const char *s);
 
-int		ft_atoi(const char *nptr);
+long	ft_atoi(const char *nptr);
 
 char	*ft_substr(char const *s, unsigned int start, size_t len);
 char	*ft_strjoin(char const *s1, char const *s2);
